Added Z3 test to check perf with simpler constraints + large numbers
Z3 finds a solution for the new test on my machine in about 1
second. The numbers I used are pretty large, which is somewhat
promising. But lots of parallelism may still be a problem.

This does sort of add evidence to the idea that if we can eliminate
the problem of irreducibility with types of form (S1 || S2) . S3,
perhaps with some Kleene-ish algebra like overapproximation like
(S1 || S2) . S3 <: (or is equivalent to?) (S1 . S3) || (S2 . S3) and
get every entailment involving || to just be at most (n1/t1 || n2/t2)
<: (n3/t3 || n4/t4), then maybe things can be much more efficient. We
would also need to allow more parallelism, of course (so not just
limited to two streams in parallel).

But I'm not sure.

Also added TeX source for some quick notes on the current approach
I've been pursuing recently. But not quite finished --- I actually
haven't gotten around to any of the more technical stuff yet --- but a
bit too tired to continue...