More work on hand-encoding SMT constraints
This is kind of breaking my brain and my fingers doing this by hand. I
do wonder if there's a nicer abstraction (i.e. a helper function of
some kind) that can automate the generation of parts of these
constraints. Specifically, the logic to convert rates to different
window sizes (either to a supertype or a subtype) is common, so you
should be able to factor it out.