Attempt fix that handles the LConcat case
The current thought for the LConcat case is basically to turn it
into a triple of conjunctions (one for left, one for crossover, one
for right). I think I can handle this without further mess by just
extending the return value of rate_symbolize to be a vector of
symbolic rates, where the conjunction is implicit.

That said, if we change the rules a bit for LConcat, things could get
pretty messy. It's already going to be a bunch of messy-ish
code. There may be a way to better modularize things and make the code
more agnostic to specific rules, but let's just punt on that for now.