Allow parsing to handle single raw rates with no parentheses
Some truly cursed, entangled code here. Lots of mutually recursive
function (that aren't explicitly declared as such, since I don't think
Rust supports/requires that) and logic in match cases that depend on
behavior of the calling function.

Would be nice to abstract things a bit better and disentangle the code
a bit, but maybe that just means learning parsing theory and
implementing an actual parser (and not whatever hacked mess this is).