Fix (I think) a mistake in SMT encoding
Also added back some debugging printout code (temporarily). The
mistake, I believe, is with the window size range constraints, which I
initially claimed would be between the minimum and maximum "seen"
window sizes during our recursive constraint construction.

This doesn't really work in the presence of concatenation, I think,
since there is then the possibility of windows that are sums between
two concrete windows, a concrete window and a symbolic window, or even
two symbolic windows. This throws a wrench in the initial constraint.

Removing the range constraints fixes at least one obvious failing
test, namely:
(. 10/5 10/5 (. 10/5 10/5 10/5) 2/1) <: 1000/3
which (I argue) should obviously be true, but which the previous
implementation was returning false for.

I think this requires a broader rethinking and actual formalization of
some of these optimization constraints that I previously added. I
would like some fully written-out proofs of why the optimizations I
add to the set of constraints actually work in all cases.

FWIW, I'm also no longer convinced that the optimization constraints
for possible window sizes (i.e. the big OR that I add at the end of
the set of constraints, restricting window size to any seen concrete
or symbolic window size) is true in all cases in the presence of
concatenation. So writing out some full proofs would be helpful in
that regard as well.