Save small comments on concrete window deduplication
Don't think it's worth doing right now, since I don't anticipate that
it will result in meaningful speedup in the general case
anyway (although it would be nice just for simplicity's sake).

I think the next pressing to-do on this implementation would be to add
back the Star operator, since Caleb and I have talked about how this
is still an actually useful construct. Now that I've mostly cobbled
together the implementation in the Concat case, the Star case should
just be a particular instance of that implementation (where the two
adjacent stream being concatenated are the same).